Finished challenge 1: roll die to determine outcome.
Defined (roll n m).
Extended execute to take the die count on both src and dst.
Extended game-tree to roll the dice and determine outcome.

I fought and fought this one. Nowhere made sense to put the code. I
think this is the right way to do it. But I'm open to suggestion.