Ch 14 medium done... This one is ugly.
Changed broadcast-universe to customize each mail's payload to the
recipient. This meant inlining and unfactoring broadcast.
serialize-universe now takes both the play and the id of the recipient
and sanitizes the waypoints for everyone else.

I'm hoping to change it so broadcast takes a function that provides
both the play and player getting the message.